make changes suggested offline
